backend/requirements.txt
# FastAPI Backend for RAG Chat
# Python 3.10+ required

fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
pandas==2.2.0

# Vector store
faiss-cpu==1.7.4
# Or use: chromadb==0.4.22

# Embeddings
sentence-transformers==2.3.1

# LLM - Hugging Face Transformers
transformers==4.37.2
torch==2.2.0
accelerate==0.26.1

# For Llama-2 access
huggingface-hub==0.20.3
